Time: 8 ms (63.85%) | Memory: 32 MB (54.02%) - LeetSync

+++ 561-array-partition/array-partition.cpp
@@ -3,7 +3,7 @@ public:
     int arrayPairSum(vector<int>& nums) {
         int n=nums.size();
         sort (nums.begin(),nums.end());
-        long long total=0;
+        double total=0;
         for (int i=0;i<n;i+=2)
         {
             total+=nums[i];
